\name{edata}
\alias{edata}
\docType{data}
\title{
Training data
}
\description{
The training data (gse14333, gse13294) used in the paper with the 47 CRCassigner genes. 
}
\usage{data("edata")}
\format{
  A data frame with 47 observations on the following 192 variables.
}
\details{
The training data (gse14333, gse13294) used in the paper with the 47 CRCassigner genes. 
}
\examples{
data(edata)
}
\keyword{datasets}
