\name{labs}
\alias{labs}
\docType{data}
\title{
Known class labels for the training data
}
\description{
Known class labels for the training data
}
\usage{data("labs")}
\format{
  The format is:
 Factor w/ 5 levels "Enterocyte","Goblet-like",..: 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
Known class labels for the training data
}
\examples{
data(labs)
## maybe str(labs) ; plot(labs) ...
}
\keyword{datasets}
