\name{pamfn}
\alias{pamfn}
\title{
Function for running pam}
\description{
Function for running pam}
\usage{
pamfn(i, p, LS.pam1, data)
}
\arguments{
  \item{i}{
Index for identifying gene sets}
  \item{p}{
Size of each gene set
}
  \item{LS.pam1}{
PAM trained model}
  \item{data}{
PAM dataset}
}
\details{
Function for running pam}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
