\name{plotMCR}
\alias{plotMCR}
\title{
Plotting MCR}
\description{
Plotting MCR}
\usage{
plotMCR(colg, ps.mean, ps.std, nrmethcl, nrmeth, p, np, nsamp)
}
\arguments{
  \item{colg}{
Colors}
  \item{ps.mean}{
Average MCR}
  \item{ps.std}{
Standard deviation of MCR}
  \item{nrmethcl}{
Number of class prediction methods}
  \item{nrmeth}{
Number of gene selection methods}
  \item{p}{
Size of each gene set}
  \item{np}{
Number of gene sets}
  \item{nsamp}{
Number of sampling}
}
\details{
Plotting MCR}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
