\name{psfn}
\alias{psfn}
\title{
Function for running pam}
\description{
Function for running pam}
\usage{
psfn(i, p, LS.resp, TrLS1, ngenes, nsubs)
}
\arguments{
  \item{i}{
Index for identifying gene sets}
  \item{p}{
Size of each gene set
}
  \item{LS.resp}{
Known classes}
  \item{TrLS1}{
Training data}
  \item{ngenes}{
Total number of genes}
  \item{nsubs}{
Number of subtypes}
}
\details{
Function for running pam}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
