\name{splitData}
\alias{splitData}
\title{
Split data into training and testing set.
}
\description{
Split data into training and testing set.
}
\usage{
splitData(response, array, fac)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{response}{
Sample class labels.
}
  \item{array}{
Sample expression data
}
  \item{fac}{
Fraction for splitting the data into training and testing set.
}
}
\details{
Split data into training and testing set.
}
\value{
\item{Training data }{Training set}
\item{Testing data }{Testing set}
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
